feat(#3): NumberInput 컴포넌트를 Input으로 변경
