--- src/components/atoms/NumberInput/NumberInput.stories.tsx
@@ -1,14 +1,0 @@
-import NumberInput from './';
-import { ComponentStory } from '@storybook/react';
-
-export default {
-  title: 'Atoms/NumberInput',
-  component: NumberInput,
-}
-
-const Template: ComponentStory<typeof NumberInput> = (args) => <NumberInput {...args} />
-
-export const Default = Template.bind({});
-Default.args = {
-  value: 0,
-}
